chore(secret-retriever): move it to opamp
