+++ agent-control/src/opamp/secret_retriever.rs
