+++ agent-control/src/opamp/secret_retriever/on_host.rs
